\name{bin2dec }
\alias{bin2dec}
\title{ Binary Number to Decimal number Conversion
}
\description{ A function returns decimal number from a sequential binary number. The function is internally used in lbp function.
}
\usage{
bin2dec(x) 
}
\arguments{
  \item{x}{
A sequence of 0, 1 numbers
}
 }
\value{
A decimal number
}
\seealso{ dec2bin, lbp
}
