\name{lbnum}
\alias{lbnum}
\title{Counts 0-1 or 1-0 in a Binary Sequence
}
\description{ A function returns how many 0-1 or 1-0 transitions in a binary sequence.  For example, 00010000 is 2 transition and 01010100 is 6 transitions. It is internally used in the lbp function.
}
\usage{
lbnum(seq)
}
\arguments{
  \item{seq}{ A sequential 0, 1 array 
}
}
\seealso{
lbp
}
